refactor(dids): extract DidKeyUtils and DidIonUtils into dedicated modules
Extract utility classes from did-key.ts and did-ion.ts into separate
focused modules to reduce file size and improve code organization:

- did-key.ts (967 -> 714 lines): DidKeyUtils -> did-key-utils.ts (258 lines)
- did-ion.ts (895 -> 718 lines): DidIonUtils -> did-ion-utils.ts (186 lines)

Both original files re-export the extracted classes for full backward
compatibility. All 302 dids tests pass, lint and build clean.
